refactor: split ArticleRankings into 2 components ------------------- In order to have standalone logic for the ranking
